partially organize components folder and update refs
