+++ client/src/components/spots/AddSpotInfo.js
