+++ client/src/components/spots/CameraButtons.js
