+++ client/src/components/spots/CategoryModal.js
